refactor: icon 이동
