+++ frontend/techpick/src/shared/const/mockdata.ts
