+++ frontend/techpick/src/shared/icons/index.ts
